refactor : make reduxAction , resetUserInfo
