+++ src/redux/reducers/AuthReducer.ts
@@ -11,29 +11,6 @@ const AuthInitialState = {
   expiredTime: null,
 };
 
-export const login_sucess = (data: any) => ({
-  //redux의 action 향후 분리필요
-  type: GET_AUTH,
-  data,
-});
-
-export const reset_auth = () => ({
-  //redux의 action 향후 분리필요
-  type: RESET_AUTH,
-});
-
-export const renew_auth = (data: any) => ({
-  //redux의 action 향후 분리필요
-  type: REFRESH_AUTH,
-  data,
-});
-
-export const get_email = (data: any) => ({
-  //redux의 action 향후 분리필요
-  type: GET_EMAIL,
-  data,
-});
-
 export const AuthReducer = (state = AuthInitialState, action: any) => {
   switch (action.type) {
     case GET_AUTH:
