+++ src/redux/reducers/UserInfoReducer.ts
@@ -1,4 +1,4 @@
-import { GET_EMAIL } from "../../utils/types";
+import { GET_EMAIL, RESET_USER_INFO } from "../../utils/types";
 const UserInfoInitialState = {
   email: null,
 };
@@ -10,6 +10,9 @@ export const UserInfoReducer = (state = UserInfoInitialState, action: any) => {
         ...state,
         email: action.data,
       };
+
+    case RESET_USER_INFO:
+      return { state: UserInfoInitialState };
 
     default:
       return state;
